MyProject.Test/Application.Test/Services.Test/PermissionGroupServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using Moq;
using MyProject.Application.Services;
using MyProject.Domain.Entities;
using MyProject.Domain.Interfaces;
using Xunit;

namespace MyProject.Tests.Services
{
    public class PermissionGroupServiceTests
    {
        [Theory]
        [InlineData(1)]
        [InlineData(2)]
        public async Task GetByIdAsync_ExistingId_ReturnsPermissionGroup(int permissionGroupId)
        {
            // Arrange
            var mockRepository = new Mock<IPermissionGroupRepository>();
            var service = new PermissionGroupService(mockRepository.Object);

            var expectedPermissionGroup = new PermissionGroup { Id = permissionGroupId, Name = "TestPermissionGroup" };
            mockRepository.Setup(repo => repo.GetByIdAsync(permissionGroupId)).ReturnsAsync(expectedPermissionGroup);

            // Act
            var result = await service.GetByIdAsync(permissionGroupId);

            // Assert
            Assert.NotNull(result);
            Assert.Equal(permissionGroupId, result.Id);
            // Add more specific assertions based on your expected behavior
        }

        [Fact]
        public async Task GetAllAsync_ReturnsAllPermissionGroups()
        {
            // Arrange
            var mockRepository = new Mock<IPermissionGroupRepository>();
            var service = new PermissionGroupService(mockRepository.Object);

            var expectedPermissionGroups = new List<PermissionGroup>
            {
                new PermissionGroup { Id = 1, Name = "PermissionGroup 1" },
                new PermissionGroup { Id = 2, Name = "PermissionGroup 2" },
                // Add more sample permission groups as needed
            };
            mockRepository.Setup(repo => repo.GetAllAsync()).ReturnsAsync(expectedPermissionGroups);

            // Act
            var result = await service.GetAllAsync();

            // Assert
            Assert.NotNull(result);
            // Add more specific assertions based on your expected behavior
        }

        [Fact]
        public async Task CreateAsync_ValidPermissionGroup_ReturnsPermissionGroupId()
        {
            // Arrange
            var mockRepository = new Mock<IPermissionGroupRepository>();
            var service = new PermissionGroupService(mockRepository.Object);

            var newPermissionGroup = new PermissionGroup { Name = "New PermissionGroup" }; // Replace with test data
            var createdPermissionGroupId = 1; // Replace with an expected ID for testing purposes
            mockRepository.Setup(repo => repo.AddAsync(newPermissionGroup)).ReturnsAsync(new PermissionGroup { Id = createdPermissionGroupId });

            // Act
            var result = await service.CreateAsync(newPermissionGroup);

            // Assert
            Assert.Equal(createdPermissionGroupId, result);
            // Add more specific assertions based on your expected behavior
        }

        // Write similar tests for UpdateAsync and DeleteAsync methods

        // Remember to add specific assertions to validate the responses and behaviors of each method.
    }
}
